fixing logic, change qty & price to floating point. add transaction data

+++ src/main/java/com/github/BamsGianYagami/POSSpringWeb/Entity/Stock.java
@@ -20,7 +20,7 @@ public class Stock {
     @GeneratedValue(strategy = GenerationType.SEQUENCE, generator = "StockitemIdGenerator")
     int itemId;
     String itemName;
-    Integer itemPrice;
-    Float qty;
+    float itemPrice;
+    float qty;
     String unitCount;
 }
